Solved

+++ problem1064.java
@@ -8,7 +8,7 @@ public class problem1064 {
             for(int i=0;i<array.length;i++){
                 array[i]=scan.nextDouble();
             }
-            double count=0;
+            int count=0;
             double sum =0;
             for(int i=0;i<array.length;i++){
                 if(array[i]>0){
